requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Principal
WTForms
Werkzeug
python-dotenv
Jinja2

spacy
pandas
mail-parser
validators
pyyaml

jsonschema
json-log-formatter

requests

tenacity
filelock

python-multipart

bcrypt
passlib[bcrypt]
itsdangerous
python-jose[cryptography]

SQLAlchemy
psycopg2-binary

APScheduler

nltk
torch
transformers
langchain
spacy-transformers

thefuzz  # Added

python-dateutil  # Added

phonenumbers  # Added

pytest
pytest-mock
unittest2
flake8
black
isort
flake8-black
sphinx

tqdm

numpy
